Added net6.0 target. Bumped BenchmarkDotNet version to 0.13.1 (#50)
