+++ src/Hashids.net/Hashids.net.csproj
@@ -13,7 +13,7 @@
         <PackageTags>ids;hash</PackageTags>
         <RepositoryType>git</RepositoryType>
         <RepositoryUrl>https://github.com/ullmark/hashids.net</RepositoryUrl>
-        <TargetFrameworks>net461;net5.0;netstandard2.0</TargetFrameworks>
+        <TargetFrameworks>net461;netstandard2.0;net5.0;net6.0</TargetFrameworks>
         <OutputType>Library</OutputType>
         <LangVersion>9</LangVersion>
         <IncludeSymbols>true</IncludeSymbols>
