+++ test/Hashids.net.test/Hashids.net.test.csproj
@@ -2,10 +2,10 @@
 
     <PropertyGroup>
 		<!--
-		The main project targets net461, net5.0, and netstandard2.0.
+		The main project targets net461, netstandard2.0, net5.0 and net6.0
 		As test projects cannot target netstandard the net48 target is used as a stand-in for that.
 		-->
-        <TargetFrameworks>net461;net48;net5.0</TargetFrameworks>
+        <TargetFrameworks>net461;net48;net5.0;net6.0</TargetFrameworks>
         <IsPackable>false</IsPackable>
         <LangVersion>9</LangVersion>
     </PropertyGroup>
